logger: move logger to utils
