+++ common/network/services/BUILD
@@ -21,10 +21,10 @@ cc_library(
     hdrs = ["udp_receiver.hpp"],
     deps = [
         "//ciri:node_shared",
-        "//common:logger_helper",
         "//common/network:neighbor",
         "//common/network/components:receiver_shared",
         "//common/network/services:receiver_shared",
+        "//utils:logger_helper",
         "@boost//:asio",
     ],
 )
@@ -35,10 +35,10 @@ cc_library(
     hdrs = ["tcp_receiver.hpp"],
     deps = [
         "//ciri:node_shared",
-        "//common:logger_helper",
         "//common/network:neighbor",
         "//common/network/components:receiver_shared",
         "//common/network/services:receiver_shared",
+        "//utils:logger_helper",
         "@boost//:asio",
     ],
 )
@@ -48,9 +48,9 @@ cc_library(
     srcs = ["tcp_sender.cc"],
     hdrs = ["tcp_sender.hpp"],
     deps = [
-        "//common:logger_helper",
         "//common/network:endpoint",
         "//common/trinary:trit_array",
+        "//utils:logger_helper",
         "@boost//:asio",
     ],
 )
@@ -60,9 +60,9 @@ cc_library(
     srcs = ["udp_sender.cc"],
     hdrs = ["udp_sender.hpp"],
     deps = [
-        "//common:logger_helper",
         "//common/network:endpoint",
         "//common/trinary:trit_array",
+        "//utils:logger_helper",
         "@boost//:asio",
     ],
 )
