+++ common/network/services/receiver.cc
@@ -5,9 +5,9 @@
  * Refer to the LICENSE file for licensing information
  */
 
-#include "common/logger_helper.h"
 #include "common/network/services/tcp_receiver.hpp"
 #include "common/network/services/udp_receiver.hpp"
+#include "utils/logger_helper.h"
 
 static char const receiver_service_logger[] = "receiver_service";
