+++ common/network/services/tcp_receiver.cc
@@ -7,8 +7,8 @@
 
 #include "common/network/services/tcp_receiver.hpp"
 #include "ciri/node.h"
-#include "common/logger_helper.h"
 #include "common/network/neighbor.h"
+#include "utils/logger_helper.h"
 
 static char const tcp_receiver_service_logger[] = "tcp_receiver_service";
