+++ common/storage/sql/sqlite3/storage.c
@@ -10,10 +10,10 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include "common/logger_helper.h"
 #include "common/storage/sql/defs.h"
 #include "common/storage/sql/statements.h"
 #include "common/storage/storage.h"
+#include "utils/logger_helper.h"
 
 #include <logger.h>
 #include <sqlite3.h>
