+++ utils/logger_helper.c
